+++ hazelcast/src/main/java/com/hazelcast/internal/partition/ReplicaFragmentMigrationState.java
@@ -18,11 +18,9 @@ package com.hazelcast.internal.partition;
 
 import com.hazelcast.cluster.Address;
 import com.hazelcast.internal.cluster.Versions;
-import com.hazelcast.internal.nio.BufferObjectDataOutput;
 import com.hazelcast.internal.partition.impl.PartitionDataSerializerHook;
 import com.hazelcast.internal.services.ServiceNamespace;
 import com.hazelcast.logging.ILogger;
-import com.hazelcast.memory.MemorySize;
 import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.nio.serialization.IdentifiedDataSerializable;
@@ -35,9 +33,8 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.function.BooleanSupplier;
 
-import static java.lang.String.format;
+import static com.hazelcast.internal.partition.ChunkSerDerHelper.readChunkedOperations;
 
 /**
  * Contains fragment namespaces along with their partition versions and migration data operations
@@ -50,10 +47,7 @@ public class ReplicaFragmentMigrationState
     private Map<ServiceNamespace, long[]> namespaces;
     private Collection<Operation> migrationOperations;
 
-    private transient Collection<ChunkSupplier> chunkSuppliers;
-    private transient int maxTotalChunkedDataInBytes;
-    private transient ILogger logger;
-    private transient int partitionId;
+    private transient ChunkSerDerHelper chunkSerDerHelper;
 
     public ReplicaFragmentMigrationState() {
     }
@@ -61,16 +55,12 @@ public class ReplicaFragmentMigrationState
     public ReplicaFragmentMigrationState(Map<ServiceNamespace, long[]> namespaces,
                                          Collection<Operation> migrationOperations,
                                          Collection<ChunkSupplier> chunkSuppliers,
-                                         int maxTotalChunkedDataInBytes, ILogger logger, int partitionId) {
-        assert chunkSuppliers != null;
-        assert logger != null;
-
+                                         int maxTotalChunkedDataInBytes, ILogger logger,
+                                         int partitionId) {
         this.namespaces = namespaces;
         this.migrationOperations = migrationOperations;
-        this.chunkSuppliers = chunkSuppliers;
-        this.maxTotalChunkedDataInBytes = maxTotalChunkedDataInBytes;
-        this.logger = logger;
-        this.partitionId = partitionId;
+        this.chunkSerDerHelper = new ChunkSerDerHelper(logger, partitionId,
+                chunkSuppliers, maxTotalChunkedDataInBytes);
     }
 
     public Map<ServiceNamespace, long[]> getNamespaceVersionMap() {
@@ -106,103 +96,7 @@ public class ReplicaFragmentMigrationState
 
         // RU_COMPAT 5.0
         if (out.getVersion().isGreaterOrEqual(Versions.V5_1)) {
-            writeChunkedOperations(out);
-        }
-    }
-
-    private void writeChunkedOperations(ObjectDataOutput out) throws IOException {
-        IsEndOfChunk isEndOfChunk = new IsEndOfChunk(out, maxTotalChunkedDataInBytes);
-
-        for (ChunkSupplier chunkSupplier : chunkSuppliers) {
-
-            chunkSupplier.signalEndOfChunkWith(isEndOfChunk);
-
-            while (chunkSupplier.hasNext()) {
-                Operation chunk = chunkSupplier.next();
-                if (chunk == null) {
-                    break;
-                }
-
-                logCurrentChunk(chunkSupplier);
-
-                out.writeObject(chunk);
-
-                if (isEndOfChunk.getAsBoolean()) {
-                    break;
-                }
-            }
-
-            if (isEndOfChunk.getAsBoolean()) {
-                logEndOfChunk(isEndOfChunk);
-                break;
-            }
-        }
-        // indicates end of chunked state
-        out.writeObject(null);
-
-        logEndOfAllChunks(isEndOfChunk);
-    }
-
-    private void logCurrentChunk(ChunkSupplier chunkSupplier) {
-        if (!logger.isFinestEnabled()) {
-            return;
-        }
-
-        logger.finest(String.format("Current chunk [partitionId:%d, %s]",
-                partitionId, chunkSupplier));
-    }
-
-    private void logEndOfChunk(IsEndOfChunk isEndOfChunk) {
-        if (!logger.isFinestEnabled()) {
-            return;
-        }
-
-        logger.finest(format("Chunk is full [partitionId:%d, maxChunkSize:%s, actualChunkSize:%s]",
-                partitionId,
-                MemorySize.toPrettyString(maxTotalChunkedDataInBytes),
-                MemorySize.toPrettyString(isEndOfChunk.bytesWrittenSoFar())));
-    }
-
-    private void logEndOfAllChunks(IsEndOfChunk isEndOfChunk) {
-        if (!logger.isFinestEnabled()) {
-            return;
-        }
-
-        boolean allDone = true;
-        for (ChunkSupplier chunkSupplier : chunkSuppliers) {
-            if (chunkSupplier.hasNext()) {
-                allDone = false;
-                break;
-            }
-        }
-
-        if (allDone) {
-            logger.finest(format("Last chunk was sent [partitionId:%d, maxChunkSize:%s, actualChunkSize:%s]",
-                    partitionId,
-                    MemorySize.toPrettyString(maxTotalChunkedDataInBytes),
-                    MemorySize.toPrettyString(isEndOfChunk.bytesWrittenSoFar())));
-        }
-    }
-
-    private static final class IsEndOfChunk implements BooleanSupplier {
-
-        private final int positionStart;
-        private final int maxTotalChunkedDataInBytes;
-        private final BufferObjectDataOutput out;
-
-        private IsEndOfChunk(ObjectDataOutput out, int maxTotalChunkedDataInBytes) {
-            this.out = ((BufferObjectDataOutput) out);
-            this.positionStart = ((BufferObjectDataOutput) out).position();
-            this.maxTotalChunkedDataInBytes = maxTotalChunkedDataInBytes;
-        }
-
-        @Override
-        public boolean getAsBoolean() {
-            return bytesWrittenSoFar() >= maxTotalChunkedDataInBytes;
-        }
-
-        public int bytesWrittenSoFar() {
-            return out.position() - positionStart;
+            chunkSerDerHelper.writeChunkedOperations(out);
         }
     }
 
@@ -224,18 +118,8 @@ public class ReplicaFragmentMigrationState
 
         // RU_COMPAT 5.0
         if (in.getVersion().isGreaterOrEqual(Versions.V5_1)) {
-            readChunkedOperations(in);
+            migrationOperations = readChunkedOperations(in, migrationOperations);
         }
-    }
-
-    private void readChunkedOperations(ObjectDataInput in) throws IOException {
-        do {
-            Object operation = in.readObject();
-            if (operation == null) {
-                break;
-            }
-            migrationOperations.add(((Operation) operation));
-        } while (true);
     }
 
     @Override
